pyproject.toml
[project]
name = "podcast-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "nemo-toolkit[asr]>=2.1.0",
    "torchaudio>=2.5.1",
]
